+++ src/noahtheduke/splint.clj
@@ -106,6 +106,7 @@
    noahtheduke.splint.rules.style.prefer-clj-math
    noahtheduke.splint.rules.style.prefer-clj-string
    noahtheduke.splint.rules.style.prefer-condp
+   noahtheduke.splint.rules.style.prefer-for-with-literals
    noahtheduke.splint.rules.style.prefer-vary-meta
    noahtheduke.splint.rules.style.reduce-str
    noahtheduke.splint.rules.style.redundant-let